Add a `--quiet` flag, inverse of `--verbose`
`--quiet` is supported as a method for decrementing verbosity, in
contrast to `-v/--verbose` incrementing it.

Because this is added as a universal option, and some Search commands
already have a `-q` option for query strings, only `--quiet` is added
for now, leaving `-q` usage available as a future enhancement.

In support of this, the controls for logging and warning verbosity are
moved into the CommandState object, and `--verbose/--quiet` counting
options invoke `set_verbosity` to tune it up or down appropriately.
`--debug` is made almost into an alias of `-vvv`, except that it also
sets `debug=True` on the state object. This ensures that
`--debug --verbose` as a combination won't result in *lower* verbosity
than `--debug`, and allows for usages like `--debug --quiet --quiet`
for more fine-grained control if necessary.