Add CLI003 and CLI004 to internal flake8 plugin
These rules try to enforce that short_help for commands always ends
with a period.

CLI003 targets function docstrings which are a single line, and
therefore likely get picked up as short-help. It impacts all
functions in the CLI, including those which are not commands, but
this keeps the implementation simple.

CLI004 targets explicit `short_help` passed in the arguments to a
method call. When it is a string or f-string, it is checked to see if
it ends in a `.`. If a call or variable name is used, no guarantee is
offered.

i.e. The following code fails CLI004:

    @command("foo", short_help="foos a bar")
    def ...

but this similar block passes CLI004:

    short_help = "foos a bar"

    @command("foo", short_help=short_help)
    def ...

Allowing for such loopholes is necessary in order to keep evaluation
of CLI004 tractable.