Refactor to scope the lifecycle of SQLite storage
The storage adapter is here rewritten entirely away from a "pure
functional style" in which attributes and values live in global scope
into a more object-oriented style. The reason for this refactor is
that we want to better control the lifecycle of the storage and ensure
that it is closed cleanly when the program exits; furthermore, we want
to simulate that same open/close behavior in the testsuite.

The overall tack taken here is
- wrap SQLiteAdapter in a new container type, CLITokenstorage
- give CLITokenstorage as many passthrough methods as are necessary to
  make this transition easy as a rewrite for various usage sites
- take steps to rewrite usages which do not have a handle on a
  LoginManager, and use the LoginManager as a way of getting access to
  the token storage
- start work on testsutie fixes

This approach has proven more successful than alternatives which were
attempted. At time of writing, `mypy` is passing on the result, some
interactive testing seems to work, and tests are in need of numerous
adjustments, but not so severe as some other failed rewrites:

  50 failed, 977 passed, 13 errors

Therefore, there should be at most 63 remediation steps, after which
this should be a functioning option for us.