Move test dependencies to a dependency group
We had dependency data for tests in a `test` extra, which is not
intended to be part of our public package interface. Move it to a
dependency group and remove the `test` extra.

Update usage of `.[test]` in the `Makefile`, which appears to be
highly outdated but which could be in use for some developers.

The `test` group had an entry for `setuptools` which appears to be
unnecessary, so it is removed here.

The dependency update script (`make update-dependencies`) gets a small
update to find where `mypy` is currently pinned.