Move 'test-mindeps' to a dependency group
Of necessity, this change also updates the min bound for `requests`
(for simplicity, to the latest version). Moving these data into a
dependency group revealed that our bound for requests was not
effective because it was in conflict with our version of `responses`.
With `deps` installed before the package and test dependencies, we
were actually not using the old version of `requests` which was
declared.

This change fixes the issue for the most part. It's still notable that
package install and dependency group installation are separate steps
in tox and can conflict.