Improve the implementation of tox `-localsdk`
It does not appear to be necessary to pass `GLOBUS_SDK_PATH` to a
subprocess, which won't work under `tox_uv` anyway (because the
virutalenv won't have `pip`).
Most likely, this method of passing is something which didn't work
in tox v3 and which v4 fixed. (It seems unlikely that we wouldn't have
tried this config.) In any case, the new simpler config works fine.