Improve error messaging on bad client creds (#1085)
* Improve error messaging on bad client creds

When the CLI is used with client creds and the creds themselves are
invalid, Auth replies with an error, and the prior enforcement logic
treated this as a case to prompt for login.

Now, we check if we know that the client creds env vars are set and
validate and message about those vars based on what's observed.

Additionally, the non-client-creds case is reworded slightly to match
the way that `globus session` prompts are provided to users.

* Refine error message emission

Improvements to internal documentation and readability.