Make the `globus api timer` command display "Globus Timers"
The goal of this change is to make the `globus api timer` command
show "Globus Timers" instead of "Globus Timer" in its help text.

The word "Timers" is computed from the text in `ServiceNameLiteral`.
Changing this ripples through the codebase, and ultimately
would modify the `globus api timer` subcommand as well.

To prevent the `globus api timer` subcommand from also changing,
the service_name literal is conditionally made singular
in a variable named `legacy_name` in `api.py`.