add support for activity-notifications for GCP and GCS
This changeset implements a syntax based on the GCS CLI's
approach to configuring guest collection activity
notification policies. Guest collection create and update
operations accept a new command line option:

--activity-notifications

To enable notifications under any circumstance, simply
set the value of the option to all. Specifying an empty
string will disable all notifications.

Finer grained policies may be specified through a syntax
that employs comma-separated strings corresponding to
the role of the collection in the transfer (transfer_use)
and outcome of the transfer (status).

The valid values for status are "succeeded" and "failed".
For transfer_use, "source" and "destination" are accepted.

The syntax builds up the policy structure that looks like
this:

{
    "status": [(string), (string)...],
    "transfer_use": [(string), (string)...]
}

Specifying a policy value for one of the attributes while
leaving out values of the others will implicitly enable
all other values from the other attribute. For example,
if "failed" is given, this is equal to the policy:

{
    "status": ["FAILED"],
    "transfer_use": ["source", "destination"]
}

Specifying one value for each attribute enables only
those situations. Fore example, specifying the value
"destination,succeeded" is equivalent to:

{
    "status": ["SUCCEEDED"],
    "transfer_use": ["destination"]
}

The syntax above is implemented through two click.ParamType subclasses:

GCSManagerGuestActivityNotificationParamType
TransferGuestActivityNotificationParamType

Two classes exist because of some differences between the GCS Manager
API and the Transfer API.  GCS collection document schema 1.14.0
implements the status values as all lowercase, whereas Transfer
expresses them as all capitals because that is how they are represented
elsewhere in the API.

Because of the way the code for the GCS commands weed out null
values from payloads, the convert() methods emit different structures
for the "" string case. For the GCS class, the structure looks like:

{
    "status": [],
    "transfer_use": []
}

whereas the Transfer class yields an ExplictNullType instance.

Both routes rely on a common decorator (activity_notifications_option),
which takes a string value of "GCS" or "GCP" to determine whether
the data type class associated with the option will be the GCS or
Transfer variant.

implement shell completion for --activity-notifications

Code borrows heavily (if not shamelessly) from the NotifyParamType.

Added changelog fragment for --activity-notifications.

Also incorporates a raft of changes based on helpful feedback
from @sirosen in pull request #1105, for which the author is grateful.