Fix ordering of task list date filters (#1100)
In #767 this ordering was "fixed" incorrectly. The "after" filter
should come first and the "before" filter should come second, so that
the date filters are formulated as

    completion_time:<after>,<before>
    request_time:<after>,<before>

The reason being, each of these forms a bound on the date range.
This was confirmed in the Get Task List documentation.[^1]

Revert that incorrect change and update the relevant test to assert
the correct ordering.

Resolves #1099

[^1]: https://docs.globus.org/api/transfer/task/#get_task_list