Update Choice subtypes to handle click compat
In order to work under both the latest and older versions of click,
custom Choice subtypes need some specialized handling.

Choice is now a generic in the type of its input and output (and
therefore, by contract, requires those two to be the same type).

In one case, the handling is simple, if verbose:
Use `click.Choice[str]` for type checking, and `click.Choice` at
runtime.

For ColonDelimitedChoiceTuple, there is no type-safe way to rewrite
this to fit the declared contract around click.Choice, so instead we
now *wrap* click.Choice in that param type, forwarding most of the
important methods to it and using them to implement our custom
`convert()` method. Method forwarding also needs to account for
compatibility and signature changes.