Fix an annotation for `make_parser()`
The click internal optparse parser is now semi-hidden: the name is
changed to `_OptionParser` and docs say "you shouldn't need to use it
in most cases". This leaves it unclear whether or not we're still okay
to keep using it, but for now, the simplest fix is to avoid explicitly
annotating the type, since `make_parser`'s return type is sufficient
for type inference to do the right things here.