Fix test suite compatibility with click 8.2.1
click 8.2.1 has a test framework change that raises `EOFError`.

It is considered a bug fix for click, but the exception was uncaught
and caused test suite failures in the CLI.

A comment regarding the usage of `.readlines()` instead of
-- presumably -- writing `enumerate(stream)` is honored
by calling `.read()` and then splitting the lines.

However, the comment itself is removed. It was written in 2016,
when Python 3.5 was the latest version of Python,
and may not be relevant any longer.