Fix `filter_roles` for `globus flows list` (#1117)
This was being passed with a default of `()`, the empty tuple. The SDK
translates this to an empty string, which the service rejects.

As a quick/immediate fix, make it `or None` so that no parameter will
be included in requests.

It was not possible to set `default=None` in the option declaration
for `filter_roles` because of the way `click` treats `multiple=True`
options -- the default is `()` even if you set an explicit default of
`None`.