Update the lower-bound for build-system.requires
resolves #1121

The lower bounded versions of our build requirements need to support
all of the features which we're using.

In #1111 we switched to PEP 639 license metadata, but we didn't update
the lower bound for `setuptools`. This closes that gap -- setuptools
v77.0.1 is the earliest available version with SPDX license expression
field support.