Address a CI caching issue
`pyproject.toml` needs to be targeted for cache busting,
not `setup.py`, which is almost completely empty.

The practical effect of this is that the cache isn't busted
and tox environments are restored successfully.
However, tox then needs to recreate the test environments
because the SDK version no longer matches what's installed
in the cached tox environments, which is a waste of CI time.