Add a script for confirming min python is tested
Based on the version of this script in `globus-sdk`, this change is a
preparatory move for removal of Python 3.8 support. When we do this,
we want to make sure we keep the minimal version of Python
well-aligned in the repo.

Use `mddj` to read the minimum version of Python from package
metadata and the minimum listed version in tox environments.
Use YAML parsing to pull the relevant data from GitHub workflows.

Add a tox environment + a CI check which runs it to call the script.

In dependency-groups data, `yaml` is now a group so that we can keep a
single pinned version for both testsuite usage and the new
`check-project-metadata` group. This group now also supplies the
dependencies for the `build`+`twine check` tox environment.