Introduce an 'ActivityFormatter' for Timers
The ActivityFormatter takes two fields from a Timer document, the
`activity` and `next_run`, and formats the results into a string.

To pass data from parsing to rendering, an auxiliary datatype is defined
as part of the formatter: ParsedActivityInfo.

For datetime formatting, we use the existing date formatter, which
localizes times. This also handles malformed data gracefully with a
warning, as confirmed in the unit tests.