Update to support click>=8.3,<8.4
This requires a newer version of `click-type-test` to support the
internal `UNSET` sentinel in `click`.

No changes to the CLI itself appear to be needed.

No changelog should be needed, as we are maintaining a bound by click
feature release, here updated to 8.4. This means that the last changelog
(which did not mention the exact version used in our bounds) remains
accurate in describing coarse-grained changes to the project.